% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00roxygen.R, R/S4methods.R
\name{ADconstruct}
\alias{ADconstruct}
\alias{diag,advector,ANY,ANY-method}
\alias{matrix,advector-method}
\title{AD aware numeric constructors}
\usage{
\S4method{diag}{advector,ANY,ANY}(x, nrow, ncol)

\S4method{matrix}{advector}(data = NA, nrow = 1, ncol = 1, byrow = FALSE, dimnames = NULL)
}
\description{
These base functions, that construct numeric like objects, have been modified to return AD objects when used inside an active AD context. This is necessary due to limitations in R's method dispatch machinery (\code{[<-} cannot dispatch on the \code{value} argument).
}
\section{Functions}{
\itemize{
\item \code{diag(x = advector, nrow = ANY, ncol = ANY)}: Equivalent of \link[base]{diag}

\item \code{matrix(advector)}: Equivalent of \link[base]{matrix}

}}
\examples{
func <- function(t) {
  x <- numeric(5)
  print(class(x))
  t
}
invisible(func(1))            ## 'numeric'
invisible(MakeTape(func, 1))  ## 'advector'
}
