% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\name{ADoverloads}
\alias{ADoverloads}
\title{Enable extra RTMB convenience methods}
\usage{
ADoverloads(enable = TRUE, ...)
}
\arguments{
\item{enable}{Logical; enable if TRUE, disable if FALSE.}

\item{...}{Passed to \link[base]{attach}, so can contain e.g. \code{warn=FALSE}.}
}
\description{
Enable extra RTMB convenience methods
}
\details{
Work around limitations in R's method dispatch system by overloading some selected primitives, currently:
\itemize{
\item Inplace replacement, so you can do \code{x[i] <- y} when \code{x} is numeric and \code{y} is AD.
\item Mixed combine, so you can do e.g. \code{c(1., x)} when \code{x} is AD.
\item Diagonal assignment, so you can do \code{diag(x) <- y} when \code{x} is a numeric matrix and \code{y} is AD.
}
}
\examples{
test <- function() {
  x <- 1:3
  y <- advector(1.2)
  x[2] <- y
  x
}
ADoverloads(TRUE) ## Enable
test() ## With ADoverloads
ADoverloads(FALSE) ## Disable
test() ## Without ADoverloads it doesn't work
}
