% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot.R
\name{ADuniroot}
\alias{ADuniroot}
\alias{uniroot,ANY-method}
\title{AD one-dimensional root finding.}
\usage{
\S4method{uniroot}{ANY}(
  f,
  interval,
  ...,
  lower = min(interval),
  upper = max(interval),
  f.lower = f(lower, ...),
  f.upper = f(upper, ...),
  extendInt = c("no", "yes", "downX", "upX"),
  check.conv = FALSE,
  tol = .Machine$double.eps^0.25,
  maxiter = 1000,
  trace = 0
)
}
\arguments{
\item{f, interval, ..., lower, upper, f.lower, f.upper, extendInt, check.conv, tol, maxiter, trace}{See \link[stats]{uniroot}.}
}
\value{
List with component \code{"root"}.
}
\description{
Univariate root finding extending R's native \link[stats]{uniroot} function to work in both \emph{standard} and \emph{AD} evaluation modes.
}
