% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLa4a-package.R
\docType{package}
\name{FLa4a-package}
\alias{FLa4a}
\alias{FLa4a-package}
\title{FLa4a: a simple and robust statistical catch-at-age model}
\description{
Minimal implementation of the a4a stock assessment model using 'RTMB'.
The main function is \code{\link[=sca]{sca()}}.
}
\author{
\strong{Maintainer}: Ernesto Jardim \email{ernesto.jardim@gmail.com}

Authors:
\itemize{
  \item Colin P. Millar \email{colin.millar@ices.dk}
}

}
\keyword{internal}
