% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00roxygen.R
\docType{package}
\name{RTMB-package}
\alias{RTMB-package}
\alias{RTMB}
\title{RTMB: R bindings for 'TMB'}
\description{
The package 'RTMB' provides a native R interface for *a subset of*
'TMB' so you can avoid coding in C++.  'RTMB' only affects the
'TMB' function 'MakeADFun' that builds the objective function. Once
'MakeADFun' has been invoked, everything else is \emph{exactly the same}
and \emph{models run as fast} as if coded in C++.
}
\author{
Kasper Kristensen

Maintainer: <kaskr@dtu.dk>
}
