% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00roxygen.R, R/complex.R
\name{Tape}
\alias{Tape}
\alias{MakeTape}
\alias{$.Tape}
\alias{TapeConfig}
\title{The AD tape}
\usage{
MakeTape(f, x)

\method{$}{Tape}(x, name)

TapeConfig(
  comparison = c("forbid", "tape", "allow"),
  atomic = c("enable", "disable"),
  vectorize = c("disable", "enable")
)
}
\arguments{
\item{f}{R function}

\item{x}{numeric vector}

\item{name}{Name of a tape method}

\item{comparison}{Set behaviour of AD comparison (\code{">"},\code{"=="}, etc).}

\item{atomic}{Set behaviour of AD BLAS operations (notably matrix multiply).}

\item{vectorize}{Enable/disable AD vectorized 'Ops' and 'Math'.}
}
\description{
The AD tape as an R function
}
\details{
A 'Tape' is a representation of a function that accepts \emph{fixed size} numeric input and returns \emph{fixed size} numeric output.
The tape can be constructed using \code{F <- MakeTape(f, x)} where \code{f} is a standard \emph{differentiable} R function (or more precisely: One using only functions that are documented to work for AD types).
Having constructed a tape F, a number of methods are available:

Evaluation:
\itemize{
\item Normal function evaluation 'F(x)' for numeric input.
\item AD evaluation 'F(x)' as part of other tapes.
\item Jacobian calculations using 'F$jacobian(x)'.
}

Transformation:
\itemize{
\item Jacobian function transformation 'F$jacfun()'
\item Laplace approximation transformation 'F$laplace()'. Also allows Saddle Point Approximation (SPA).
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{$}: Get a tape method.

}}
\section{Functions}{
\itemize{
\item \code{MakeTape()}: Generate a 'Tape' of an R function.

\item \code{TapeConfig()}: Global configuration parameters of the tape (experts only!)
\bold{comparison} By default, AD comparison gives an error
(\code{comparison="forbid"}).
This is the safe and recommended behaviour, because comparison is a
non-differentiable operation. If you are building a tape that
requires indicator functions e.g. \code{f(x)*(x<0)+g(x)*(x>=0)}
then use \code{comparison="tape"} to add the indicators to the
tape. A final option \code{comparison="allow"} exists for
testing/illustration purposes. Do not use.

}}
\examples{
F <- MakeTape(prod, numeric(3))
show(F)
F$print()
H <- F$jacfun()$jacfun() ## Hessian tape
show(H)
\dontrun{
#### Handy way to plot the graph of F
if (require(igraph)) {
   G <- graph_from_adjacency_matrix(F$graph())
   plot(G, vertex.size=17, layout=layout_as_tree)
}
}
}
