% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\name{ADoverloads}
\alias{ADoverloads}
\title{Enable extra RTMB convenience methods}
\usage{
ADoverloads()
}
\description{
Enable extra RTMB convenience methods
}
\details{
Work around limitations in R's method dispatch system by overloading some selected primitives, currently:
\itemize{
\item Inplace replacement, so you can do \code{x[i] <- y} when \code{x} is numeric and \code{y} is AD.
\item Mixed combine, so you can do e.g. \code{c(x, y)} when \code{x} numeric and \code{y} is AD.
\item Diagonal assignment, so you can do \code{diag(x) <- y} when \code{x} is a numeric matrix and \code{y} is AD.
}

In all cases, the result should be AD.
The methods are automatically \strong{temporarily} attached to the search path (\code{search()}) when entering \link{MakeTape} or \link{MakeADFun}.
Alternatively, by running \code{ADoverloads()}, the methods will be added to the calling environment (\code{parent.frame()}). This is only necessary when using RTMB from a package.
}
\examples{
MakeTape(function(x) {print(search()); x}, numeric(0))
MakeTape(function(x) c(1,x), 1:3)
MakeTape(function(x) {y <- 1:3; y[2] <- x; y}, 1)
MakeTape(function(x) {y <- matrix(0,3,3); diag(y) <- x; y}, 1:3)
}
