% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpol.R
\name{Interpolation}
\alias{Interpolation}
\alias{interpol1Dfun}
\alias{interpol2Dfun}
\title{Interpolation}
\usage{
interpol1Dfun(z, xlim = c(1, length(z)), ...)

interpol2Dfun(z, xlim = c(1, nrow(z)), ylim = c(1, ncol(z)), ...)
}
\arguments{
\item{xlim}{Domain of x}

\item{...}{Configuration parameters}

\item{ylim}{Domain of y}
}
\value{
function of x.

function of x and y.
}
\description{
Some interpolation methods are available to be used as part of 'RTMB' objective functions.
}
\section{Functions}{
\itemize{
\item \code{interpol1Dfun()}: Construct a kernel smoothed representation of a vector.

\item \code{interpol2Dfun()}: Construct a kernel smoothed representation of a matrix.

}}
\examples{
f <- interpol2Dfun(volcano, xlim=c(0,1), ylim=c(0,1))
F <- MakeTape(function(x) f(x[1],x[2]), c(.5,.5))
}
