% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00roxygen.R, R/complex.R
\name{Tape}
\alias{Tape}
\alias{MakeTape}
\alias{$.Tape}
\title{The AD tape}
\usage{
MakeTape(f, x)

\method{$}{Tape}(x, name)
}
\description{
The AD tape as an R function
}
\details{
A 'Tape' is a representation of a function that accepts \emph{fixed size} numeric input and returns \emph{fixed size} numeric output.
The tape can be constructed using \code{F <- MakeTape(f, x)} where \code{f} is a standard \emph{differentiable} R function (or more precisely: One using only functions that are documented to work for AD types).
Having constructed a tape F, a number of methods are available:

Evaluation:
- Normal function evaluation 'F(x)' for numeric input.
- AD evaluation 'F(x)' as part of other tapes.
- Jacobian calculations using 'F$jacobian(x)'.

Transformation:
- Jacobian function transformation 'F$jacfun()'
- Laplace approximation transformation 'F$laplace()'. Also allows Saddle Point Approximation (SPA).
}
\section{Methods (by generic)}{
\itemize{
\item \code{$}: Get a tape method.

}}
\section{Functions}{
\itemize{
\item \code{MakeTape()}: Generate a 'Tape' of an R function.

}}
\examples{
F <- MakeTape(prod, numeric(3))
show(F)
F$print()
H <- F$jacfun()$jacfun() ## Hessian tape
show(H)
}
